fix!: add wakemeops prefix to repo url
This change makes the configuration of apt-cacher-ng a lot easier and is
a good practice of debian repositories.

The new repo url is: http://deb.wakemeops.com/wakemeops/.

The old url is still supported using a redirection rule.

As a result the repo Origin and Label have changed and users need to
manually accept those changes.